move make_results helper modules to results_processing
